Moved ManualProjectCreatorDialogTest back to collect_app module
