+++ collect_app/src/main/res/layout/manual_project_creator_dialog_layout.xml
